+++ packages/blade/src/components/Alert/Alert.stories.tsx
@@ -79,7 +79,6 @@ const meta: Meta<AlertProps> = {
     isDismissable: true,
     contrast: 'low',
     intent: 'information',
-    isBorderless: false,
     actions: {
       primary: {
         text: 'Primary Action',
@@ -189,7 +188,15 @@ PrimaryActionOnly.parameters = {
   },
 };
 
-export const FullWidth = AlertTemplate.bind({});
+export const FullWidth: ComponentStory<typeof AlertComponent> = ({ ...args }) => {
+  return (
+    <Box height={200} position="relative">
+      <Box position="absolute" width="100%">
+        <AlertComponent {...args} />
+      </Box>
+    </Box>
+  );
+};
 FullWidth.args = {
   description: 'Currently you can only accept payments in international currencies using PayPal.',
   intent: 'notice',
@@ -200,12 +207,13 @@ FullWidth.args = {
 FullWidth.parameters = {
   docs: {
     description: {
-      story: 'A full width Alert can be used to span the entire width of its container',
+      story:
+        'A full width Alert can be used to span the entire width of its container. It also makes the Alert borderless and can be used for full-bleed layouts. You can also wrap the alert and adjust layout with absolute positioning if needed.',
     },
   },
 };
 
-export const Borderless: ComponentStory<typeof AlertComponent> = ({ ...args }) => {
+export const FullWidthWithActions: ComponentStory<typeof AlertComponent> = ({ ...args }) => {
   return (
     <Box height={200} position="relative">
       <Box position="absolute" width="100%">
@@ -214,17 +222,16 @@ export const Borderless: ComponentStory<typeof AlertComponent> = ({ ...args }) =
     </Box>
   );
 };
-Borderless.args = {
-  description: 'Use vendor payouts to quickly generate invoices.',
-  actions: undefined,
-  title: undefined,
-  isBorderless: true,
+FullWidthWithActions.args = {
+  description: 'Currently you can only accept payments in international currencies using PayPal.',
+  intent: 'negative',
+  isFullWidth: true,
 };
-Borderless.parameters = {
+FullWidthWithActions.parameters = {
   docs: {
     description: {
       story:
-        'A borderless Alert can be used for full-bleed layouts, this automatically makes the alert full width. You can also wrap the alert and adjust layout with absolute positioning if needed.',
+        'A full width Alert with `actions` will render them inline if there is enough space and responsively wrap them to the next line in smaller displays.',
     },
   },
 };