+++ packages/blade/src/components/Alert/__tests__/Alert.web.test.tsx
@@ -25,18 +25,6 @@ describe('<Alert />', () => {
     expect(container).toMatchSnapshot();
   });
 
-  it('should render negative intent and borderless', () => {
-    const { container } = renderWithTheme(
-      <Alert
-        description="Currently you can only accept payments in international currencies using PayPal."
-        intent="negative"
-        isBorderless
-      />,
-    );
-
-    expect(container).toMatchSnapshot();
-  });
-
   it('should handle onClick on actions', async () => {
     const user = userEvent.setup();
     const onClickPrimary = jest.fn();